requirements.txt
boto3==1.34.69
python-dotenv==1.0.1
